+++ src/main.cpp
@@ -2,148 +2,134 @@
 #include "allocator.h"
 #include "garbage_collector.h"
 
+// Example Class to showcase Allocator and GC
+class MyClass {
+public:
+	int myint;
+
+	MyClass(int myint): myint(myint) {
+		std::cout << "My Class Constructor Called" << std::endl;
+		std::cout << "My Class ID = " << myint << std::endl;
+	}
+
+	~MyClass(){
+		std::cout << "My Class Destructor Called" << std::endl;
+		std::cout << "My Class ID = " << myint << std::endl;
+	}
+
+	void foo() {
+		std::cout << "My Class Function Called" << std::endl;
+		std::cout << "My Class ID = " << myint << std::endl;
+	}
+
+};
+
 int main(){
-	bool DEBUG_MODE = true;
-	Allocator& alloc = Allocator::getInstance(DEBUG_MODE);
+	
+	// By Default, DEBUG_MODE = false. To enable debug logs, you can do DEBUG_MODE = true
+	/*
+		bool DEBUG_MODE = true;
+		Allocator& alloc = Allocator::getInstance(DEBUG_MODE);		
+	*/
+
+	Allocator& alloc = Allocator::getInstance();
+
+	// By Default, GC_ENABLED = true. You can automatic garbage collection by setting GC_ENABLED = false;
+
+	alloc.GC_ENABLED = true;
+
+	// If you want to manually invoke GC, you can get the instance of GC from the allocator.
+	
 	Garbage_Collector& gc = alloc.getGC();
 
-	int* arr = (int*)alloc.allocate(10 * sizeof(int), (void**) &arr);
+	// Let us create an array of size 3 with MyClass Objects
 
-	std::cout << "&arr = " << &arr << std::endl;	
-	std::cout << "arr = " << arr << std::endl;		
-
-	for (int i = 0; i < 10; i++) {
-		arr[i] = i;
-	}
-
-	for (int i = 0; i < 10; i++) {
-		std::cout << arr[i] << " ";
-	}												// 0 1 2 3 4 5 6 7 8 9
-
-	std::cout << std::endl;
-	
-	gc.gc_dump();
-	gc.gc_collect();
-	alloc.heap_dump();
-
-	int** arr2 = (int**)alloc.allocate(3 * sizeof(int*), (void**)&arr2);
-	std::cout << "&arr2 = " << &arr2 << std::endl;
-	std::cout << "arr2 = " << arr2 << std::endl;
+	MyClass** arr = (MyClass**)alloc.allocate(3*sizeof(MyClass*), (void**) & arr);
 
 	for (int i = 0; i < 3; i++) {
-		arr2[i] = (int*)alloc.allocate(10 * sizeof(int));
-		for (int j = 0; j < 10; j++) {
-			arr2[i][j] = (i + 1) * j;
-		}
+		// The first argument of allocate_new is the stack variable reference which provides root for GC
+		// If provided, the GC will not collect the area referenced by the stack variable during garbage collection
+		// You can decide to not provide it by setting the 1st argument = nullptr. As long as the parent node is referencing the heap, the gc will not collect any of the child nodes
+		// The next 'n' arguments are the arguments for the constructor. The arguments are forwarded to the constructor
+		// Here there is only 1 var in Constructor, id. So we will provide the id
+		int myObjId = i;
+		arr[i] = alloc.allocate_new<MyClass>(nullptr, myObjId);
 	}
-	
-	std::cout << "&arr2 = " << &arr2 << std::endl;			// 0x7ffc26e13ba0
-	std::cout << "arr2 = " << arr2 << std::endl;			// 0x5603f55b2478
+	/*
+	OUTPUT:
+	My Class Constructor Called
+	My Class ID = 0
 
+	My Class Constructor Called
+	My Class ID = 1
+	
+	My Class Constructor Called
+	My Class ID = 2
+	*/
+
+
+
+	// Let us invoke the foo() function
 	for (int i = 0; i < 3; i++) {
-		std::cout << arr2[i] << " : " << std::endl << "\t";
-		for (int j = 0; j < 10; j++) {
-			std::cout << arr2[i][j] << ", ";
-		}
-		std::cout << std::endl;
+		arr[i]->foo();
 	}
 
-	alloc.heap_dump();
-	gc.gc_dump();
-	gc.gc_collect();
-	alloc.heap_dump();
+	/*
+	OUTPUT:
+	My Class Function Called
+	My Class ID = 0
 	
-	/*
-				--- SNIPPET OF DEBUG LOGS ---
-	Chunks:
-		Chunk at: 0x5603f55b2400, Size: 40 bytes, Allocated, gc_mark : MARKED, Next: 0x5603f55b2450, Prev: 0
-		Chunk at: 0x5603f55b2450, Size: 24 bytes, Allocated, gc_mark : UNMARKED, Next: 0x5603f55b2490, Prev: 0x5603f55b2400
-		Chunk at: 0x5603f55b2490, Size: 40 bytes, Allocated, gc_mark : UNMARKED, Next: 0x5603f55b24e0, Prev: 0x5603f55b2450
-		Chunk at: 0x5603f55b24e0, Size: 40 bytes, Allocated, gc_mark : UNMARKED, Next: 0x5603f55b2530, Prev: 0x5603f55b2490
-		Chunk at: 0x5603f55b2530, Size: 40 bytes, Allocated, gc_mark : UNMARKED, Next: 0, Prev: 0x5603f55b24e0
-	*/
-
-
-
-	// assign function ensures that that the elements are not picked by gc during sweeping phase
-	// it marks the area 'reachable'
-
-	int* a = alloc.assign(&a, arr2[0]);
-	int* b = alloc.assign(&b, arr2[1]);
-	int* c = alloc.assign(&c, arr2[2]);
-
-	/*
-			--- SNIPPET OF DEBUG LOGS ---[
-	Chunks:
-		Chunk at: 0x5603f55b2400, Size: 40 bytes, Allocated, gc_mark : MARKED, Next: 0x5603f55b2450, Prev: 0
-		Chunk at: 0x5603f55b2450, Size: 24 bytes, Allocated, gc_mark : MARKED, Next: 0x5603f55b2490, Prev: 0x5603f55b2400
-		Chunk at: 0x5603f55b2490, Size: 40 bytes, Allocated, gc_mark : MARKED, Next: 0x5603f55b24e0, Prev: 0x5603f55b2450
-		Chunk at: 0x5603f55b24e0, Size: 40 bytes, Allocated, gc_mark : MARKED, Next: 0x5603f55b2530, Prev: 0x5603f55b2490
-		Chunk at: 0x5603f55b2530, Size: 40 bytes, Allocated, gc_mark : MARKED, Next: 0, Prev: 0x5603f55b24e0
-	*/
-
-
-	std::cout << "------- SETTING arr2 = nullptr -------" << std::endl;
-
-	arr2 = nullptr;
-
-	gc.gc_dump();
-	gc.gc_collect();
-	alloc.heap_dump();
-
-	/*
-				--- SNIPPET OF DEBUG LOGS ---
-	Chunks:
-		Chunk at: 0x5603f55b2400, Size: 40 bytes, Allocated, gc_mark : MARKED, Next: 0x5603f55b2450, Prev: 0
-		Chunk at: 0x5603f55b2450, Size: 24 bytes, Free, gc_mark : UNMARKED, Next: 0x5603f55b2490, Prev: 0x5603f55b2400
-		Chunk at: 0x5603f55b2490, Size: 40 bytes, Allocated, gc_mark : MARKED, Next: 0x5603f55b24e0, Prev: 0x5603f55b2450
-		Chunk at: 0x5603f55b24e0, Size: 40 bytes, Allocated, gc_mark : MARKED, Next: 0x5603f55b2530, Prev: 0x5603f55b2490
-		Chunk at: 0x5603f55b2530, Size: 40 bytes, Allocated, gc_mark : MARKED, Next: 0, Prev: 0x5603f55b24e0
+	My Class Function Called
+	My Class ID = 1
+	
+	My Class Function Called
+	My Class ID = 2
 	
 	*/
 
-
-	std::cout << "------- SETTING a,b,c = nullptr -------" << std::endl;
-
-	a = nullptr;
-	b = nullptr;
-	c = nullptr;
-
-	gc.gc_dump();
-	gc.gc_collect();
-	alloc.heap_dump();
-
+	// Let us look at snippet of heap
+	alloc.heap_dump();		// Only works if DEBUG_MODE = true
 	/*
-		
-			--- SNIPPET OF DEBUG LOGS ---
-
-	[INFO]    Starting sweeping phase..
-
-	[INFO]          Sweeping pointer -> 0x5603f55b2450
-
-	[INFO]          Sweeping pointer -> 0x5603f55b2490
-
-	[INFO]          Coalescing with previous chunk -> 0x5603f55b2450
-
-	[INFO]          Sweeping pointer -> 0x5603f55b2450
-
-	[INFO]          Sweeping pointer -> 0x5603f55b24e0
-
-	[INFO]          Coalescing with previous chunk -> 0x5603f55b2450
-
-	[INFO]          Sweeping pointer -> 0x5603f55b2450
-
-	[INFO]          Sweeping pointer -> 0x5603f55b2530
-
-	[INFO]          Coalescing with previous chunk -> 0x5603f55b2450
-
-	[INFO]          Sweeping pointer -> 0x5603f55b2450
-
-	[INFO]    Finished Sweeping phase
-
 	Chunks:
-		Chunk at: 0x5603f55b2400, Size: 40 bytes, Allocated, gc_mark : MARKED, Next: 0x5603f55b2450, Prev: 0
-		Chunk at: 0x5603f55b2450, Size: 264 bytes, Free, gc_mark : UNMARKED, Next: 0, Prev: 0x5603f55b2400
+		Chunk at: 0x5587e0d5e400, Size: 24 bytes, Allocated, gc_mark : UNMARKED, Next: 0x5587e0d5e440, Prev: 0
+		Chunk at: 0x5587e0d5e440, Size: 4 bytes, Allocated, gc_mark : UNMARKED, Next: 0x5587e0d5e46c, Prev: 0x5587e0d5e400
+		Chunk at: 0x5587e0d5e46c, Size: 4 bytes, Allocated, gc_mark : UNMARKED, Next: 0x5587e0d5e498, Prev: 0x5587e0d5e440
+		Chunk at: 0x5587e0d5e498, Size: 4 bytes, Allocated, gc_mark : UNMARKED, Next: 0, Prev: 0x5587e0d5e46c
 	*/
+
+	// Now if we want to assign the memory to variable, you can do so using the 'assign' function
+	// Assign function ensures that the garbage collector will not pick up the objects during the collection process
+
+	MyClass* ptr = alloc.assign(&ptr, arr[2]);
+
+	// Now if we set the arr to nullptr or some other variable, it will not collect the area pointed by obj
+	arr = nullptr;
+
+	// Let us manually invoke the GC
+	gc.gc_collect();
+	alloc.heap_dump();		// Only works if DEBUG_MODE = true
+	/*
+	Chunks:
+		Chunk at: 0x559db2cbd400, Size: 112 bytes, Free, gc_mark : UNMARKED, Next: 0x559db2cbd498, Prev: 0
+		Chunk at: 0x559db2cbd498, Size: 4 bytes, Allocated, gc_mark : MARKED, Next: 0, Prev: 0x559db2cbd400
+	*/
+	// As you can see, all the unreachable chunks have been collected by the gc
+
+	// Important Note: Garbage collector does not invoke the destructor of the object
+	// To safely destroy an object. Use the free_ptr() function
+
+	alloc.free_ptr<MyClass>(ptr);
+	/*
+	My Class Destructor Called
+	My Class ID = 2
+	*/
+
+	alloc.heap_dump();		// Only works if DEBUG_MODE = true
+	/*
+	Chunks:
+		Chunk at: 0x559db2cbd400, Size: 156 bytes, Free, gc_mark : UNMARKED, Next: 0, Prev: 0
+	*/
+
+
 
 }
